Adds new DSL for models
Modifies the model class to remove some crud for now. There is a new
meta API coming in the next version of dry struct that we can use
instead of the hacky way we have dealt with the new, saved and parent
props now.
Refactors the types to not monkey patch dry-types for #is support and
adds a new collector to collect all of the input from the DSL block
before finalising the class.
Fixes some spec errors due to the above changes, commented out some as
well since I'm not gonna solve them right now.
Adds exemplar files to have something to exercise the DSL on. I'm going
to implement part of the https://reqres.in/ mock API as tests for now.